Project specific deploy keys
Because github won't allow the same deploy key to be used in multiple
projects, each project has to have its own deploy key. Right now the
implementation is a bit messy because I could not get docker insert to
work, so we're just building a Dockerfile on the fly and build a new
image with the keys + ssh/config inside it.