server-side project discovery
Project discovery now happens server side. This will avoid hitting CORS
problem with github, as well as crippling the console.

All relevant code has been moved to a Discoverer\Github service which is
in turn called in the ProjectController (and in the
user:project:discover command, yay for reusability of services \o/).